Fixed/updated frequency options
Also switched back to using default picker style as section headers don't work properly in the navigation link style.

+++ Phonepedia/Model/SwiftData/Phone.swift
@@ -125,7 +125,7 @@ final class Phone {
 	
 	var hasTransmitOnlyBase: Bool = false
 	
-	var frequency: Int = 1920
+    var frequency: Double = 1920
     
     var hasNoLineAlert: Bool = false
